SignUpViewControllerでメアド&パスを使って簡単なJSONを取得

+++ iOS/app/app/Talk.h
@@ -11,10 +11,10 @@
 @interface Talk : NSObject
 // インスタンス変数
 @property (nonatomic, copy) NSString *name;
-@property (nonatomic, assign) int periodStart;
-@property (nonatomic, assign) int periodEnd;
-@property (nonatomic, assign) int member;
-@property (nonatomic, assign) int icon;
-@property (nonatomic, assign) int posts;
+@property (nonatomic, assign) NSInteger periodStart;
+@property (nonatomic, assign) NSInteger periodEnd;
+@property (nonatomic, assign) NSInteger member;
+@property (nonatomic, assign) NSInteger icon;
+@property (nonatomic, assign) NSInteger posts;
 
 @end
